# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductController_index",
        "parameters": [
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "sku",
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": 0,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "offset",
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "sort_by",
            "in": "query",
            "schema": {
              "default": "created_at",
              "type": "string"
            }
          },
          {
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "required": false,
            "name": "sort_dir",
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductIndexResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Indexar productos",
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductController_store",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductStoreRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductStoreResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "409": {
            "description": "El sku ya esta registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear producto",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductController_show",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductShowResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mostrar producto",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequestDocDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductUpdateResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "409": {
            "description": "El sku ya esta registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar producto",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductController_destroy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDestroyResponseDocDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request invalido: body, query o params no cumplen el contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationErrorResponseDocDto"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token requerido, invalido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "403": {
            "description": "El usuario autenticado no tiene el rol requerido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          },
          "500": {
            "description": "Error inesperado del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDocDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar producto",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bus Impl API",
    "description": "API de negocio actualizada a 2026",
    "version": "2026.06",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ApiValidationErrorResponseDocDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Campo del request que fallo la validacion."
                },
                "message": {
                  "type": "string",
                  "description": "Mensaje legible del error de validacion."
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "ProductIndexResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "sku": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 40
                },
                "nombre": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 120
                },
                "descripcion": {
                  "anyOf": [
                    {
                      "type": "string",
                      "maxLength": 500
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "precio": {
                  "type": "string",
                  "pattern": "^\\d{1,10}(\\.\\d{1,2})?$"
                },
                "stock": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "active",
                    "inactive",
                    "archived"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "sku",
                "nombre",
                "descripcion",
                "precio",
                "stock",
                "status",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            },
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "count",
          "next",
          "previous",
          "results"
        ]
      },
      "ProductShowResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "sku": {
                "type": "string",
                "minLength": 2,
                "maxLength": 40
              },
              "nombre": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "descripcion": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 500
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "precio": {
                "type": "string",
                "pattern": "^\\d{1,10}(\\.\\d{1,2})?$"
              },
              "stock": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "archived"
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "sku",
              "nombre",
              "descripcion",
              "precio",
              "stock",
              "status",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "product"
        ]
      },
      "ProductStoreRequestDocDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "minLength": 2,
            "maxLength": 40,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "precio": {
            "type": "string",
            "pattern": "^\\d{1,10}(\\.\\d{1,2})?$",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "sku",
          "nombre",
          "descripcion",
          "precio",
          "stock"
        ]
      },
      "ProductStoreResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "product": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "sku": {
                "type": "string",
                "minLength": 2,
                "maxLength": 40
              },
              "nombre": {
                "type": "string",
                "minLength": 2,
                "maxLength": 120
              },
              "descripcion": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 500
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "precio": {
                "type": "string",
                "pattern": "^\\d{1,10}(\\.\\d{1,2})?$"
              },
              "stock": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "archived"
                ]
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              }
            },
            "required": [
              "id",
              "sku",
              "nombre",
              "descripcion",
              "precio",
              "stock",
              "status",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-self-required": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "product"
        ]
      },
      "ProductUpdateRequestDocDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "minLength": 2,
            "maxLength": 40,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "precio": {
            "type": "string",
            "pattern": "^\\d{1,10}(\\.\\d{1,2})?$",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "stock": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "archived"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        }
      },
...
      "ProductDestroyResponseDocDto_Output": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "message": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}